Refactored part of genome editor
