+++ core/src/main/java/io/github/some_example_name/old/editor/di/EditorVariables.kt
@@ -1,0 +1,12 @@
+package io.github.some_example_name.old.editor.di
+
+interface EditorVariables {
+    var currentTick: Int
+    var currentStage: Int
+    var lastTick: Int
+    var lastStage: Int
+    var grabbedCellIndex: Int
+    var lastGrabbedCellX: Float
+    var lastGrabbedCellY: Float
+    var isRightClick: Boolean
+}
