+++ core/src/main/java/io/github/some_example_name/old/editor/entities/EditorReplay.kt
@@ -1,0 +1,6 @@
+package io.github.some_example_name.old.editor.entities
+
+interface EditorReplay {
+    fun reset()
+    fun copy()
+}
