+++ core/src/main/java/io/github/some_example_name/old/editor/system/logic/LastActionType.kt
@@ -1,0 +1,5 @@
+package io.github.some_example_name.old.editor.system.logic
+
+enum class LastActionType  {
+    DIVIDE, MUTATE, CHANGE, DELETE
+}
