+++ core/src/main/java/io/github/some_example_name/old/editor/system/logic/RestartSimulationCallBack.kt
@@ -1,0 +1,5 @@
+package io.github.some_example_name.old.editor.system.logic
+
+interface RestartSimulationCallBack {
+    fun restartSimulation()
+}
