+++ core/src/main/java/io/github/some_example_name/old/editor/system/logic/UiEditorCommands.kt
@@ -1,4 +1,4 @@
-package io.github.some_example_name.old.editor.commands
+package io.github.some_example_name.old.editor.system.logic
 
 sealed interface UiEditorCommands
