+++ core/src/main/java/io/github/some_example_name/old/editor/undo_redo_commands/ChangeDivideCommand.kt
@@ -1,26 +1,20 @@
-package io.github.some_example_name.old.editor.commands
+package io.github.some_example_name.old.editor.undo_redo_commands
 
 import io.github.some_example_name.old.editor.entities.EditorCell
 import io.github.some_example_name.old.systems.genomics.genome.Action
 import io.github.some_example_name.old.systems.genomics.genome.GenomeStage
 
 class ChangeDivideCommand(
-    override val stage: Int,
     val clickedCell: EditorCell,
     val divide: Action,
-    val genomeStageInstruction: MutableList<GenomeStage>
-) : Command {
-
-    private val oldGenomeStageInstruction = genomeStageInstruction.map { it.deepCopy() }
-    private var newGenomeStageInstruction: List<GenomeStage>? = null
+    stageInstruction: MutableList<GenomeStage>,
+    currentStage: Int
+) : UndoRedoCommand(
+    stage = currentStage,
+    genomeStageInstruction = stageInstruction
+) {
 
     override fun execute() {
-        if (newGenomeStageInstruction != null) {
-            genomeStageInstruction.clear()
-            genomeStageInstruction.addAll(newGenomeStageInstruction!!)
-            return
-        }
-
         genomeStageInstruction[stage].cellActions.compute(clickedCell.parentId) { _, oldValue ->
             oldValue?.copy(divide = oldValue.divide?.copy(
                 cellType = divide.cellType,
@@ -37,12 +31,5 @@ class ChangeDivideCommand(
                 pheromoneType = divide.pheromoneType
             ))
         }
-
-        newGenomeStageInstruction = genomeStageInstruction.map { it.deepCopy() }
-    }
-
-    override fun undo() {
-        genomeStageInstruction.clear()
-        genomeStageInstruction.addAll(oldGenomeStageInstruction)
     }
 }
