+++ core/src/main/java/io/github/some_example_name/old/ui/screens/SettingsScreen.kt
@@ -257,7 +257,7 @@ object GlobalSettings {
     var DRAW_LINK_SHADER = true
     var HYDRO_ENABLED = false
     var HYDRO_VISUALIZATION = false
-    var MUSIC_VOLUME = 50
+    var MUSIC_VOLUME = 0
     var SOUND_VOLUME = 50
     var GRID_WIDTH = gridWidth
     var GRID_HEIGHT = gridHeight
